refactor: 불필요한 컴포넌트 제거
