+++ src/pages/NotFound/index.jsx
@@ -1,10 +1,12 @@
-import LinkButton from '../../components/LinkButton';
+import Button from '../../components/Button';
 
 const NotFound = () => {
   return (
     <div>
       <h1>존재하지 않는 페이지입니다 </h1>
-      <LinkButton to="/">홈으로 가기</LinkButton>
+      <Button type="link" to="/">
+        홈으로 가기
+      </Button>
     </div>
   );
 };
